Do not reset network for each sample run.
This allow to lower memory usage when wanting to output a large number of token, as this is done by smaller batch.
To allow that one new command line parameter is added `-count` that give the number of block we want to generate.
`-length` is used to give the size of a block.

To allow this, sample has been modified to not reset and reseed the network each time it is run.
And a new method to LM has been added to reset and seed the network.